organismsコンポーネント分割、CleatTimeコンポーネント作成 #19
